% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\title{File archiving}
\usage{
archive(fileName = NULL, currentDir = NULL, archiveDir = NULL,
  verbose = FALSE)
}
\arguments{
\item{fileName}{Character string containing the name of the file to be archived}

\item{currentDir}{The current directory of the file}

\item{archiveDir}{Directory to which the file should be archived}

\item{verbose}{Logical that prints date and archive related messages if TRUE}
}
\description{
This function receives a file and is able to archive it to a given directory
}

