% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeMobenzi.R
\name{decodeMobenzi}
\alias{decodeMobenzi}
\title{Decodes data from mobenzi.}
\usage{
decodeMobenzi(dfSurvey = NULL, dfCodeBook = NULL,
  fldnmVariable = "Variable", fldnmValue = "Value",
  fldnmLabel = "Label", replacementChar = "_", formatOpsies = FALSE,
  fixnames = TRUE)
}
\arguments{
\item{dfSurvey}{data.frame containing the data to be decoded.}

\item{dfCodeBook}{data.frame containing the codebook.}

\item{formatOpsies}{if true, the labels containing the options will be fixnamed.}
}
\value{
a decoded dataframe
}
\description{
Decodes data from mobenzi.
}
