% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{duplicates}
\usage{
duplicates(x)
}
\arguments{
\item{x}{[data.frame/vector] The data frame/vector containing duplicates.}
}
\value{
duplicate rows/values
}
\description{
function that identifies duplicates
}
