% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixname.R
\name{fixname}
\alias{fixname}
\title{Make variable names lower case}
\usage{
fixname(data_names)
}
\arguments{
\item{data_names}{Names to be converted (function converts data into an array)}
}
\value{
Simplified variable names
}
\description{
Receives variable names and converts them to lower 
case, without any spaces or punctuation marks - whether
within the name, leading or trailing
}

