% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFields.R
\name{format_char}
\alias{format_char}
\title{Format integer-like character field}
\usage{
format_char(field = NULL)
}
\arguments{
\item{field}{Vector to be converted to a character vector}
}
\value{
Formatted field returned with replaced values
}
\description{
Function that receives an integer-like character field that converts it to lower case and matches 
various patterns whereby it replaces them with chosen values
}

