% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFields.R
\name{format_int}
\alias{format_int}
\title{Format integer-like character field}
\usage{
format_int(field = NULL)
}
\arguments{
\item{field}{Vector to be converted to a character vector}
}
\value{
Formatted field returned as a integer vector
}
\description{
Function that receives an integer-like character field and removes leading zeros,
punctuation marks, alphabetic characters and blanks
}
