% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFields.R
\name{format_int_char}
\alias{format_int_char}
\title{Format integer-like character field}
\usage{
format_int_char(field = NULL)
}
\arguments{
\item{field}{Vector to be converted to a character vector}
}
\value{
Formatted field returned as a character vector
}
\description{
Function that receives an integer-like character field and removes leading zeros,
punctuation marks, alphabetic characters and blanks using the format.int function, but 
returns a character value
}

