% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFields.R
\name{format_num}
\alias{format_num}
\title{Format integer-like character field}
\usage{
format_num(field = NULL)
}
\arguments{
\item{field}{Vector}
}
\value{
Formatted field returned as a numeric vector
}
\description{
Function that receives an integer-like character field and removes leading zeros,
punctuation marks, alphabetic characters and blanks
}

