% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFields.R
\name{format_phonenumber}
\alias{format_phonenumber}
\title{Format integer-like character field}
\usage{
format_phonenumber(field = NULL)
}
\arguments{
\item{field}{Vector to be converted to a character vector}
}
\value{
Formatted field returned as a character vector
}
\description{
Function that receives an integer-like character field referring to a phone number and 
formats the field to remove leading zeros, punctuation marks, alphabetic characters and blanks 
as well as the 'plus sign' and extra numbers
}
