% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItEasy.R
\name{generateRandomString}
\alias{generateRandomString}
\title{Generate random string}
\usage{
generateRandomString(charSet = NULL, stringLength = 7, replace = TRUE)
}
\arguments{
\item{charSet}{Character vector}

\item{stringLength}{Length of the string to be created}

\item{replace}{Logical that determines whether sampling should be with replacement or not}
}
\value{
Returns a random string using character vector concatenation
}
\description{
Generates a random string from a set of characters
}

