% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItEasy.R
\name{getFileExtension}
\alias{getFileExtension}
\title{Get file extension}
\usage{
getFileExtension(fileName = NULL)
}
\arguments{
\item{fileName}{Character string containing the name of the file}
}
\value{
File extension
}
\description{
Get file extension
}

