% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItEasy.R
\name{is.nada}
\alias{is.nada}
\title{Search for "NADA" values}
\usage{
is.nada(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Logical specifying whether "NADA" values were found
}
\description{
Search for "NADA" values
}
