% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItEasy.R
\name{isSimilar}
\alias{isSimilar}
\title{Match fields}
\usage{
isSimilar(record1 = NULL, record2 = NULL, fieldsToMatch = NULL,
  numRequiredMatches = NULL, verbose = FALSE)
}
\arguments{
\item{record1}{Array containing data related to the first record}

\item{record2}{Array containing data related to the second record}

\item{fieldsToMatch}{Either an integer vector containing the fields to match or a character vector 
containing the indices of the fields to match}

\item{numRequiredMatches}{Integer that specifies the number of required matches in order for the
fields to be considered as matching}

\item{verbose}{Logical that displays function information when TRUE}
}
\value{
Depending on the information supplied, the function will return FALSE for no matches or when 
number of matches is less than number of required matches, or return true when the records matchs completely, 
also returning the fields that match
}
\description{
This function calculates which fields of two given objects are similar
}

