% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynnapoli.R
\name{lyneNaPoligone}
\alias{lyneNaPoligone}
\title{lyneNaPoligone}
\usage{
lyneNaPoligone(fn, druk = TRUE)
}
\arguments{
\item{fn}{Karakter. Naam van ESRI shapefile soos "erf_9382_subsections_blocks_20190821.shp"}

\item{druk}{Logies. Druk of nie.}
}
\value{
"sf" "data.frame"
}
\description{
Vat lyne getrek op QGIS en maak tot poligone
}
