% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynnapoli.R
\name{maakInvulNommers}
\alias{maakInvulNommers}
\title{maakInvulNommers}
\usage{
maakInvulNommers(ssf = ll)
}
\arguments{
\item{ssf}{sf.  produk van lyneNaPoligone}
}
\description{
Vat die produk van lyneNaPoligone en skep 'n 
karakterstring in die vorm id = "" wat jy kan gebruik om 'n 
benoemde vektor van id en teikenwaarde te maak
}
