% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.col.R
\name{move.col}
\alias{move.col}
\title{Move column}
\usage{
move.col(df = NULL, colName = NULL, colIdx = NULL)
}
\arguments{
\item{df}{Data frame containing data to be manipulated}

\item{colName}{Character string containing the name of the column to be moved}

\item{colIdx}{Integer that refers to the index that the column must be moved to}
}
\value{
Returns the data frame with the column moved to the specified index
}
\description{
This function moves a column in a given data frame to a desired index in that data frame
}

