% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItEasy.R
\name{nada}
\alias{nada}
\title{Convert NA into "NADA"}
\usage{
nada(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Vector with NA values replaced with "NADA
}
\description{
Convert NA into "NADA"
}

