% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynnapoli.R
\name{nuwePoliNommers}
\alias{nuwePoliNommers}
\title{nuwePoliNommers}
\usage{
nuwePoliNommers(ssf, vnommer, druk = TRUE)
}
\arguments{
\item{vnommer}{karakter. vektor geskep en geredigeer met behulp van maakInvulNommers]}

\item{druk}{Logies}
}
\value{
ssf
}
\description{
Vat die resultaat van maakInvulNommers en lyneNaPoligone en ken nommers toe
}
