% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.col.R
\name{remove.col}
\alias{remove.col}
\title{Remove column}
\usage{
remove.col(df = NULL, colName = NULL, colIdx = NULL)
}
\arguments{
\item{df}{Data frame containing data to be manipulated}

\item{colName}{Character string containing the name of the column to be removed}

\item{colIdx}{Integer that refers to the index of the column to be removed}
}
\value{
Returns the data frame with the specified column removed
}
\description{
This function removes a specified column from a given data frame
}
