% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.col.R
\name{remove.cols}
\alias{remove.cols}
\title{Remove columns}
\usage{
remove.cols(df = NULL, colNames = NULL, colIdxx = NULL)
}
\arguments{
\item{df}{Data frame containing data to be manipulated}

\item{colNames}{Vector containing the names of the columns to be removed}

\item{colIdxx}{Vector that refers to the indices of the columns to be removed}
}
\value{
Returns the data frame with the specified columns removed
}
\description{
This function removes any number of culumns in a given data frame
}
