% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMobenzi2.R
\name{splitMobenzi2}
\alias{splitMobenzi2}
\title{Decodes Mobenzi data with extraction of repeating sections}
\usage{
splitMobenzi2(filePaths, tidy = FALSE, twoLists = FALSE, formatOptions = FALSE)
}
\arguments{
\item{filePaths}{A character argument containing the paths to .csv files (one per 
section) as downloaded from Mobenzi.}

\item{tidy}{Should sections with the same number of rows be combined into one 
section?}

\item{twoLists}{If TRUE, returns two lists - the first containing the data 
frame(s) with the actual data and the second containing the metadata, 
question book and code book. If FALSE, returns only one list with the different
data frames (those from the data as well as the dfs for the metadata, code book
and question book) simply as separate items in the list.}

\item{formatOtions}{If true, formats the question options of code book variables
with 'format_char'.}
}
\value{
One or two lists of data frames. See params 'tidy' and 'twoLists' for
more information.
}
\description{
Decodes Mobenzi data with extraction of repeating sections
}
