% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSurnameAndInitials.R
\name{splitSurnameAndInitials}
\alias{splitSurnameAndInitials}
\title{Separates initials and surnames combined in strings.}
\usage{
splitSurnameAndInitials(x)
}
\arguments{
\item{x}{[character] A vector of strings that each contain any combination of surname and initials.}
}
\value{
A data frame with three columns: 'original', 'surname' and 'initials'.
}
\description{
Separates initials and surnames combined in strings.
}
