% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAverageX.R
\name{timeAverageX}
\alias{timeAverageX}
\title{timeAverageX}
\usage{
timeAverageX(df, interval, startDate = NA)
}
\arguments{
\item{df}{A dataframe containing a date field (named 'date') in POSIXct format
and one or more numeric fields to be averaged. Any non-numeric fields will be
dropped.}

\item{interval}{Interval to average to, in minutes}

\item{startDate}{Date from which to start averaging. All observations prior to
that date will be dropped. Argument must be in POSIXct format. If no argument
is supplied, function will default to the oldest date the 'date' field of df}
}
\description{
timeAverageX
}
