% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItEasy.R
\name{uniqueS}
\alias{uniqueS}
\title{Wrapper for 'unique function'}
\usage{
uniqueS(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector, data frame or array}

\item{na.rm}{Logical that removes NA values if TRUE}
}
\value{
Vector, data frame or array with duplicate values removed
}
\description{
Added 'na.rm' option to standard unique function
}
