% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeItEasy.R
\name{unnada}
\alias{unnada}
\title{Undo changes}
\usage{
unnada(x)
}
\arguments{
\item{x}{Vector}
}
\value{
If "NADA" values are found, they are converted back to NA
}
\description{
Undo changes
}

